requirements.txt
mcp>=1.0.0
anthropic>=0.80.0
python-dotenv>=1.0.0
mcp-server-fetch>=2025.0.0
mcp-server-sqlite>=0.6.0
fastapi>=0.115.0
uvicorn>=0.34.0
